link pod again
