+++ ios/Podfile
@@ -1,8 +1,16 @@
 # Uncomment this line to define a global platform for your project
-platform :ios, '8.0'
-# Uncomment this line if you're using Swift
-# use_frameworks!
+# platform :ios, '9.0'
 
-target "RNImagePreview" do
-  pod 'JTSImageViewController'
+target 'RNImagePreview' do
+  # Uncomment this line if you're using Swift or would like to use dynamic frameworks
+  # use_frameworks!
+
+  # Pods for RNImagePreview
+
+  target 'RNImagePreviewTests' do
+    inherit! :search_paths
+    # Pods for testing
+    pod 'JTSImageViewController'
+  end
+
 end
