+++ ios/RNImagePreview/AppDelegate.h
@@ -10,6 +10,7 @@
 #import <UIKit/UIKit.h>
 
 @interface AppDelegate : UIResponder <UIApplicationDelegate>
+
 @property (nonatomic, strong) UIWindow *window;
 
 @end
