+++ package.json
@@ -6,6 +6,7 @@
     "start": "react-native start"
   },
   "dependencies": {
-    "react-native": "^0.19.0"
+    "react": "15.2.0",
+    "react-native": "^0.29.0"
   }
 }
